Added file used for errP exp in Roma, May 2018
The included file allows to run experiments (in theory, both training and testing versions) using the errP signal, decoded online, to modulate dwell time of key selection. Training showed acceptable results on me, however there appeared to be a major problem with this implementation: subjects tend to look for the next key without waiting for feedback of selected key, thus producing error potentials at unexpected (and unpredictable) times, if at all. This was initially supposed when Giacinto evaluated the interface and definitely confirmed on patients in Rome tests. Conclusion: while the idea was cool and appears to be working on specifically tasked subjects, it requires a change in spontaneous use of the keyboard, which is exactly the opposite of the point of the whole project. errP for keyboard use will probably be put aside.